requirements.txt
streamlit==1.32.2
pandas==2.2.1
numpy==1.26.4
nltk==3.8.1
spacy==3.7.4
torch==2.2.1
transformers==4.40.1
scikit-learn==1.4.2
openpyxl==3.1.2
python-docx==1.1.0
pymupdf==1.23.26
bcrypt==4.1.2
pywin32==306
streamlit_pdf_viewer==0.1.4
